% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.zigdag.R
\name{nonlinear.zidag}
\alias{nonlinear.zidag}
\title{Learning the causal structure of a nonlinear ZiG-DAG}
\usage{
nonlinear.zidag(
  dat,
  start,
  nbasis = 4,
  maxiter = 500,
  tol = .Machine$double.eps^0.25,
  optim.control = list(),
  verbose = FALSE
)
}
\arguments{
\item{dat}{a data matrix.}

\item{start}{a square adjacency matrix, the directed acyclic graph to be used to initialize the algorithm. If none is specified, the empty graph is used.}

\item{nbasis}{an integer, the number of cubic B-spline basis functions to be used for the nonlinear ZiG-DAG.}

\item{maxiter}{an integer, the maximum number of iterations.}

\item{tol}{a numeric value, the tolerance for the convergence of the network score (BIC).}

\item{optim.control}{a list of control parameters passed to \code{optim}.}

\item{verbose}{a boolean value. If \code{TRUE}, progress of the algorithm is printed; otherwise the function is completely silent.}
}
\value{

}
\description{
\code{nonlinear.zidag} learns the causal structure of a nonlinear ZiG-DAG using a hill-climbing greedy search algorithm. Specifically, the conditional distribution of each node in the model is assumed to be a zero-inflated hyper-Poisson distribution.
}
